:truck: (interface): Move Activity interface in include folder
