+++ libs/ActivityKit/tests/mocks/Activity.h
@@ -4,8 +4,8 @@
 
 #pragma once
 
-#include "Activity.h"
 #include "gmock/gmock.h"
+#include "interface/libs/Activity.h"
 
 namespace leka::mock {
